CSI_Ingester/development/pyproject.toml
[project]
name = "csi-ingester"
version = "0.1.0"
description = "Creator Signal Intelligence ingester service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.12.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]
